added handleInput / handleOutput
