+++ src/components/Leaderboard/Referrals.tsx
@@ -6,7 +6,7 @@ import Modal from 'components/Modal'
 import { ethers } from 'ethers'
 import { useBRP, useReferralContract } from 'hooks/useContract'
 import useBlockNumber from 'lib/hooks/useBlockNumber'
-import { InputSection } from 'pages/Swap'
+import { InputSection } from 'pages/Trade'
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import { Copy } from 'react-feather'
 import { useTransactionAdder } from 'state/transactions/hooks'
