+++ src/components/PoolSimple/index.tsx
@@ -24,7 +24,7 @@ import { useLimweth, useVaultContract } from 'hooks/useContract'
 import { useOnClickOutside } from 'hooks/useOnClickOutside'
 import { useStablecoinValue } from 'hooks/useStablecoinPrice'
 import { useUSDPriceBNV2 } from 'hooks/useUSDPrice'
-import { ArrowContainer } from 'pages/Swap'
+import { ArrowContainer } from 'pages/Trade'
 import React, { ReactNode, useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import { ArrowUpRight, ChevronDown, ChevronUp, Maximize2 } from 'react-feather'
 import { Info } from 'react-feather'
