+++ src/components/PositionTable/LeveragePositionTable/DecreasePositionContent/index.tsx
@@ -34,9 +34,9 @@ import { useUSDPrice } from 'hooks/useUSDPrice'
 import JSBI from 'jsbi'
 import { formatBNToString } from 'lib/utils/formatLocaleNumber'
 import tryParseCurrencyAmount from 'lib/utils/tryParseCurrencyAmount'
-import { DynamicSection } from 'pages/Swap/tradeModal'
-import { PriceToggleSection } from 'pages/Swap/tradeModal'
-import { Filter, FilterWrapper, Selector, StyledSelectorText } from 'pages/Swap/tradeModal'
+import { DynamicSection } from 'pages/Trade/tradeModal'
+import { PriceToggleSection } from 'pages/Trade/tradeModal'
+import { Filter, FilterWrapper, Selector, StyledSelectorText } from 'pages/Trade/tradeModal'
 import { useCallback, useEffect, useMemo, useState } from 'react'
 import { AlertTriangle } from 'react-feather'
 import { parseBN } from 'state/marginTrading/hooks'
