+++ src/pages/Trade/intervals.ts
