+++ src/state/swap/SwapContext.tsx
@@ -1,0 +1,6 @@
+import { Currency, Percent } from '@uniswap/sdk-core'
+
+export interface CurrencyState {
+  inputCurrency?: Currency
+  outputCurrency?: Currency
+}
